+++ internal/controller/operator/factory/vlsingle/vlogs.go
@@ -150,7 +150,7 @@ func makeVLogsPodSpec(r *vmv1beta1.VLogs) (*corev1.PodTemplateSpec, error) {
 	// if customStorageDataPath is not empty, do not add pvc.
 	shouldAddPVC := r.Spec.StorageDataPath == ""
 
-	storagePath := vlsingleDataDir
+	storagePath := dataDataDir
 	if r.Spec.StorageDataPath != "" {
 		storagePath = r.Spec.StorageDataPath
 	}
@@ -186,14 +186,14 @@ func makeVLogsPodSpec(r *vmv1beta1.VLogs) (*corev1.PodTemplateSpec, error) {
 
 	if storageSpec == nil {
 		volumes = append(volumes, corev1.Volume{
-			Name: vlsingleDataVolumeName,
+			Name: dataVolumeName,
 			VolumeSource: corev1.VolumeSource{
 				EmptyDir: &corev1.EmptyDirVolumeSource{},
 			},
 		})
 	} else if shouldAddPVC {
 		volumes = append(volumes, corev1.Volume{
-			Name: vlsingleDataVolumeName,
+			Name: dataVolumeName,
 			VolumeSource: corev1.VolumeSource{
 				PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
 					ClaimName: r.PrefixedName(),
@@ -204,7 +204,7 @@ func makeVLogsPodSpec(r *vmv1beta1.VLogs) (*corev1.PodTemplateSpec, error) {
 	volumes = append(volumes, r.Spec.Volumes...)
 	vmMounts := []corev1.VolumeMount{
 		{
-			Name:      vlsingleDataVolumeName,
+			Name:      dataVolumeName,
 			MountPath: storagePath,
 		},
 	}