+++ test/e2e/vtsingle_test.go
@@ -145,7 +145,6 @@ var _ = Describe("test vtsingle Controller", Label("vt", "single", "vtsingle"), 
 							},
 							RetentionPeriod: "1",
 							StorageDataPath: "/custom-path/internal/dir",
-							Storage:         &corev1.PersistentVolumeClaimSpec{},
 						},
 					},
 					func(cr *vmv1.VTSingle) {
@@ -156,8 +155,8 @@ var _ = Describe("test vtsingle Controller", Label("vt", "single", "vtsingle"), 
 						Expect(ts.Containers).To(HaveLen(1))
 						Expect(ts.Volumes).To(HaveLen(2))
 						Expect(ts.Containers[0].VolumeMounts).To(HaveLen(2))
-						Expect(ts.Containers[0].VolumeMounts[0].Name).To(Equal("data"))
-						Expect(ts.Containers[0].VolumeMounts[1].Name).To(Equal("unused"))
+						Expect(ts.Containers[0].VolumeMounts[0].Name).To(Equal("unused"))
+						Expect(ts.Containers[0].VolumeMounts[1].Name).To(Equal("data"))
 
 					}),
 			)